group routes into catalog, lists, and planning
Move store, shopping list, and item match endpoints into new catalog, lists, and planning packages.
Update API router wiring and route exports to use new modules.
Remove deprecated route_plans/core init and fix tests to target planning job endpoints.
