+++ backend/app/api/routes/catalog/__init__.py
@@ -1,0 +1,6 @@
+"""Catalog domain router."""
+
+from .catalog_endpoints import router
+
+__all__ = ["router"]
+
