+++ backend/app/api/routes/catalog/catalog_endpoints.py
@@ -1,77 +1,21 @@
-"""Store endpoints."""
+"""Catalog (store) HTTP endpoints."""
 
 from __future__ import annotations
 
-from datetime import datetime
 from typing import Sequence
 from uuid import UUID
 
 from fastapi import APIRouter, Depends, Query, status
 from geoalchemy2 import Geography
-from pydantic import BaseModel, ConfigDict, Field
 from sqlalchemy import cast, func
 from sqlmodel import Session, select
 
-from backend.app.api.utils.geography import extract_point_coordinates
 from backend.app.dependencies import get_db
 from backend.app.models import Store, StoreChain
 
+from .catalog_schemas import StoreChainsWithStoresResponse, StoreResponse
+
 router = APIRouter()
-
-
-class StoreResponse(BaseModel):
-    """API response payload for a single store."""
-
-    model_config = ConfigDict(from_attributes=True)
-
-    id: UUID
-    chain_id: UUID | None
-    name: str
-    number: str
-    address_line1: str
-    address_line2: str | None = None
-    city: str
-    region: str
-    postal_code: str
-    country_code: str
-    timezone: str
-    phone: str
-    external_ref: dict | None = None
-    hours_json: dict | None = None
-    longitude: float | None = None
-    latitude: float | None = None
-    created_at: datetime
-    updated_at: datetime
-
-    @classmethod
-    def from_model(cls, store: Store, geography_geojson: str | None) -> "StoreResponse":
-        """Instantiate a response object from a SQLModel store instance."""
-
-        base_data = store.model_dump(
-            exclude={
-                "geography",
-                "chain",
-                "store_products",
-                "plan_selected",
-                "store_visits",
-                "item_matches",
-            }
-        )
-        coords = extract_point_coordinates(geography_geojson)
-        if coords is not None:
-            base_data.update({"longitude": coords[0], "latitude": coords[1]})
-
-        return cls(**base_data)
-
-
-class StoreChainsWithStoresResponse(BaseModel):
-    """Response model for store chains with stores."""
-
-    model_config = ConfigDict(from_attributes=True)
-
-    id: UUID
-    name: str
-    stores: list[StoreResponse] = Field(default_factory=list)
 
 
 @router.get(
@@ -144,3 +88,6 @@ def get_nearby_store_chains(
         chain_payload.stores.append(StoreResponse.from_model(store, geojson))
 
     return list(chain_map.values())
+
+
+__all__ = ["router"]
