+++ backend/app/api/routes/catalog/catalog_helpers.py
@@ -1,0 +1,3 @@
+"""Helper utilities for catalog routes."""
+
+__all__ = []
