+++ backend/app/api/routes/lists/__init__.py
@@ -1,0 +1,6 @@
+"""Lists domain router."""
+
+from .lists_endpoints import router
+
+__all__ = ["router"]
+
