--- backend/core/__init__.py
@@ -1,6 +1,0 @@
-"""Core domain models and services."""
-
-from .config import Settings, settings
-from .parsing import parsing 
-
-__all__ = ["Settings", "settings", "parsing"]
